Fix DateTime field width & refactoring
